# src/main/resources/custom/openapi3.json
{
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ]
      },
      "AppInformation": {
        "type": "object",
        "properties": {
          "package_name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "version_number": {
            "type": "integer"
          }
        },
        "required": [
          "package_name",
          "version",
          "version_number"
        ]
      },
      "UserInformation": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "user_email",
          "user_name"
        ]
      },
      "DeviceInformation": {
        "type": "object",
        "properties": {
          "platform_type": {
            "type": "string"
          },
          "os_name": {
            "type": "string"
          },
          "os_version": {
            "type": "string"
          },
          "info_map": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "platform_type",
          "os_name",
          "info_map"
        ]
      },
      "DisplayInformation": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string"
          },
          "left": {
            "type": "integer"
          },
          "top": {
            "type": "integer"
          },
          "right": {
            "type": "integer"
          },
          "bottom": {
            "type": "integer"
          },
          "orientation": {
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "left",
          "top",
          "right",
          "bottom",
          "orientation"
        ]
      },
      "DeviceStorageInformation": {
        "type": "object",
        "properties": {
          "storage_name": {
            "type": "string"
          },
          "free_space": {
            "type": "integer"
          },
          "total_space": {
            "type": "integer"
          }
        },
        "required": [
          "storage_name",
          "free_space",
          "total_space"
        ]
      },
      "StackTraceElem": {
        "type": "object",
        "properties": {
          "file_name": {
            "type": "string"
          },
          "class_name": {
            "type": "string"
          },
          "method_name": {
            "type": "string"
          },
          "line_number": {
            "type": "integer"
          },
          "is_native_method": {
            "type": "boolean"
          }
        },
        "required": [
          "class_name",
          "method_name",
          "line_number",
          "is_native_method"
        ]
      },
      "ThreadInformation": {
        "type": "object",
        "properties": {
          "thread_id": {
            "type": "integer"
          },
          "thread_name": {
            "type": "string"
          },
          "priority": {
            "type": "integer"
          },
          "is_alive": {
            "type": "boolean"
          },
          "thread_state": {
            "type": "string"
          },
          "is_daemon": {
            "type": "boolean"
          },
          "is_interrupted": {
            "type": "boolean"
          },
          "thread_traces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StackTraceElem"
            }
          }
        },
        "required": [
          "thread_id",
          "thread_name",
          "priority",
          "is_alive",
          "thread_state",
          "is_daemon",
          "is_interrupted",
          "thread_traces"
        ]
      },
      "DeviceState": {
        "type": "object",
        "properties": {
          "is_connected_to_internet": {
            "type": "boolean"
          },
          "current_thread": {
            "$ref": "#/components/schemas/ThreadInformation"
          },
          "other_threads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadInformation"
            }
          },
          "time_zone": {
            "type": "string"
          },
          "storage_information": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceStorageInformation"
            }
          },
          "info_map": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "current_thread",
          "storage_information"
        ]
      },
      "LogEntry": {
        "type": "object",
        "properties": {
          "device_state": {
            "$ref": "#/components/schemas/DeviceState"
          },
          "tag": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack_trace": {
            "type": "string"
          },
          "values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "time_fired": {
            "type": "integer"
          }
        },
        "required": [
          "tag",
          "type",
          "message",
          "stack_trace",
          "values",
          "time_fired"
        ]
      },
      "DatabaseLogEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceInformation"
          },
          "app_information": {
            "$ref": "#/components/schemas/AppInformation"
          },
          "user_information": {
            "$ref": "#/components/schemas/UserInformation"
          },
          "log": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogEntry"
            }
          },
          "created_at": {
            "type": "integer"
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "state"
        ]
      }
    }
  },
  "info": {
    "version": "1.0.0",
    "title": "Logging Backend Schema",
    "description": "Logging backend is the backend provider for log requests generated by mobile apps, for crashes and app analytical logging",
    "termsOfService": "https://qawaz.github.io/terms",
    "contact": {}
  },
  "paths": {
    "/v1/token":{
      "summary": "Request a bearer token",
      "description": "A bearer token is returned with expiration time",
      "post": {
        "parameters": [],
        "operationId": "generateToken",
        "responses": {
          "200": {
            "description": "A request token has been generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "request_token": {
                      "type": "string"
                    },
                    "exp_time": {
                      "type": "string"
                    },
                    "encryption_key": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "A token that already exists is invalid and user must wait",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "error creating request token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "a token couldn't be created due to an internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/logs/upload": {
      "summary": "Upload a logs file",
      "description": "Takes log file and saves it",
      "post": {
        "parameters": [],
        "operationId": "uploadLogFile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The file has been uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Bad request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Request Token couldn't be authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "File couldn't be uploaded or processed because of an internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "is_encrypted": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "is_encrypted"
                ]
              }
            }
          },
          "description": "Request body for the log file upload request , containing is_encrypted"
        }
      }
    },
    "/v1/logs/list": {
      "post": {
        "summary": "List logs for a package",
        "description": "lists the logs for a package",
        "parameters": [],
        "operationId": "listLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logs are returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DatabaseLogEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error because of bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Error , request token couldn't be authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "There was a server error listing the logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "package_name": {
                    "type": "string"
                  },
                  "page": {
                    "type": "integer"
                  }
                },
                "required": [
                  "package_name",
                  "page"
                ]
              }
            }
          },
          "description": "Lists the logs for the given package and page"
        }
      }
    }
  }
}